initial success with docs build after moving files around
