--- docs/apidocs/solve.rst
@@ -1,6 +1,0 @@
-.. _qiskit_dynamics-solve:
-
-.. automodule:: qiskit_dynamics.solve
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
