--- docs/apidocs/solver_classes.rst
@@ -1,6 +1,0 @@
-.. _qiskit_dynamics-solver:
-
-.. automodule:: qiskit_dynamics.solver_classes
-   :no-members:
-   :no-inherited-members:
-   :no-special-members:
