+++ docs/apidocs/solvers.rst
@@ -1,0 +1,6 @@
+.. _qiskit_dynamics-solvers:
+
+.. automodule:: qiskit_dynamics.solvers
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
